refactor(l5): move audit-log from src/ into server/security behind an L5 adapter (issue 864)
Moved:
  src/audit-log.ts -> server/security/audit-log.ts

src/ closure under server/main.ts: 35 -> 34.

The module's single environment read was the zero-argument default of
readMcpAuditConfig, plus the `deps.config ?? readMcpAuditConfig()` fallback in
installMcpAudit. server/ reads no environment, so the server/ copy takes those
values as one required McpAuditEnvironment parameter and treats an absent
config as "audit off" rather than reaching for process.env.

server/main.ts is the runtime importer: the import path changes and the call
passes the value it already holds, ServerConfig.mcpAudit. Its field set
(enabled, retentionDays, cleanupIntervalMs, writeTimeoutMs) matches
McpAuditConfig exactly, so no mapping is needed.

src/audit-log.ts becomes an L5 adapter rather than a shim: src/operator-doctor.ts
calls readMcpAuditConfig() with no arguments and src/tools/index.ts relies on the
absent-config environment fallback. Both call forms are preserved in the adapter,
which reads process.env at call time and delegates. Every relative import in it
names a server/ path.

The server/ copy is linted whole by the pre-commit gate, so seven pre-existing
findings in the moved code are paid off here, with behavior held fixed:
- payloadSizeBucket: the ten-arm if-ladder becomes a PAYLOAD_SIZE_BUCKETS table.
- estimateJsonBytes: the scalar arms move to estimateScalarBytes and the
  container arm to visitContainer/containerStructuralBytes/pushContainerMembers.
  Structural bytes are still counted before members are pushed, so the
  early-exit path for an adversarially wide container is unchanged.
- summarizeMcpAudit: the five caller columns move to callerIdentityFields.

Tests stay at their src/ paths per M1 EXCEPTION and exercise the moved code
through the adapter, unmodified.

Verified:
- bunx tsc --noEmit -> exit 0
- oxlint --deny-warnings on both staged server/ files -> exit 0
- bun run test:isolated over src/audit-log.test.ts,
  src/tools/__tests__/mcp-audit-log.test.ts, server/main.pg.test.ts,
  src/operator-doctor.test.ts -> 58 pass, 0 fail
- bash scripts/done-means/864-moved-out-of-src.sh -> judged=32 shims,
  1 adapters PASS
